requirements.txt
--find-links=https://girder.github.io/large_image_wheels GDAL
geemap
geopandas
jupyter-server-proxy
leafmap
nbserverproxy
owslib
folium
streamlit_folium
streamlit
pymysql
pyyaml
pandas
